Add bad content & sketch for backups
This was a hacked-together sketch of rsync-based backups.
But when it came time to test it out... I decided I didn't like it.

This involves giving the reMarkable SSH access to my NAS. That can (in
principle) be limited in various ways... but all of those ways involve
something on the "honor system", a process that _usually_ executes
subcommands deciding _not_ to.

I'd rather take the approach of using a service that by default does
_not_ execute subshells - something like restic's rest-server.