Enable restic backups to NAS; pinned versions
As with other Restic backups, the repository setup requires an `init`
step that is outside the scope of these scripts. But, seems to be
running happily via Tailscale now.

Also pin the versions of Tailscale and Restic rather than picking up
from head. Really, this is to make upgraes easier - otherwise, have to
manually mangel the repositories to get at them.