Add spinscripts - PID control
From TrueNAS forums. Same idea as fanbang, but (a) PID rather than
bang-bang control, and (b) includes disk as the primary PID control
(assumes linear for CPU, which is fine).

I haven't tried these out yet. I think I'd need to move the disks (and
controller) from a VM to the host to have them influence this loop-
which is only "PID" on the disk side.

Maybe I'll do that...sometime.